diplay user profile with edit button
