+++ src/client/index.css
@@ -1,6 +1,210 @@
+@import url('https://fonts.googleapis.com/css2?family=Roboto:wght@400;500;700&display=swap');
+@import url('https://fonts.googleapis.com/css?family=Montserrat:300,400,700&display=swap');
+
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
+
+*{
+  margin: 0;
+  padding: 0;
+  box-sizing: border-box;
+  font-family: 'Montserrat', sans-serif;
+  list-style: none;
+}
+
+.wrapper {
+  background-color: #1F2937;
+  margin-top: 90px;
+  margin-bottom: 90px;
+  position: relative;
+  left: 50%;
+  transform: translateX(-50%);
+  border-radius: 20px;
+  overflow: hidden;
+  padding: 20px;
+  display: flex;
+  justify-content: space-between;
+  width: 850px;
+}
+
+.right-container {
+  background-color: white;
+  width: 200px;
+  border-radius: 20px;
+  transform: translateX(-2%);
+  /* border: 2px solid red; */
+  padding: 20px;
+}
+
+/* .wrapper,
+.profile {
+  border: 2px solid red;
+} */
+
+.wrapper {
+  width: 650px;
+  height: auto;
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%,-50%);
+  display: flex;
+}
+
+.wrapper .profile {
+  width: 400px;
+  margin-right: 25px;
+}
+
+.wrapper .profile .profile_img_info {
+  display: flex;
+  background: #fff;
+  border-radius: 10px;
+  margin-bottom: 25px;
+  box-shadow: 0 1px 2px rgba(0,0,0,0.125);
+}
+
+.wrapper .profile .profile_img_info .img {
+  width: 125px;
+  padding: 10px;
+  background: #fff;
+  border-top-left-radius: 10px;
+  border-bottom-left-radius: 10px;
+}
+
+.wrapper .profile .profile_img_info .img img {
+  width: calc(100% - 20px);
+  display: block;
+  border-radius: 10px;
+  margin: auto;
+}
+/* 
+.profile_img_info {
+  border: 2px solid red;
+} */
+
+.wrapper .profile .profile_img_info .info {
+  width: calc(100% - 125px);
+  padding: 40px;
+}
+
+.wrapper .profile .profile_img_info .info .name {
+  font-size: 24px;
+  font-weight: 700;
+  margin-bottom: 5px;
+}
+
+.wrapper .profile .profile_img_info .place {
+  font-weight: 300;
+}
+
+.wrapper .profile .profile_skills {
+  background: #fff;
+  border-radius: 10px;
+  padding: 25px;
+  box-shadow: 0 1px 2px rgba(0,0,0,0.125);
+}
+
+/* .profile_skills {
+  border: 2px solid red;
+} */
+
+.wrapper .profile .profile_skills .skills {
+  padding-bottom: 20px;
+  border-bottom: 1px solid #cccccc;
+}
+
+.wrapper .profile .profile_skills .skills p {
+  text-transform: uppercase;
+  font-weight: 700;
+  font-size: 14px;
+  margin-bottom: 10px;
+}
+
+.wrapper .profile .profile_skills .skills ul li {
+  font-size: 30px;
+  margin-bottom: 5px;
+}
+
+.wrapper .profile .profile_skills .skills ul li .fab {
+  width: 40px;
+}
+
+.wrapper .profile .profile_skills .tags_wrap {
+  padding-top: 20px;
+}
+
+.wrapper .profile .profile_skills .tags_wrap span.tag {
+  background: #e4e9ef;
+  padding: 10px 15px;
+  display: inline-block;
+  border-radius: 10px;
+  margin: 5px;
+  font-size: 14px;
+}
+
+.wrapper .profile_counts {
+  background: #fff;
+  border-radius: 10px;
+  padding: 25px;
+  width: 175px;
+  box-shadow: 0 1px 2px rgba(0,0,0,0.125);
+}
+
+.wrapper .profile_counts .profile_counts_wrap {
+  padding: 30px 0;
+  border-bottom: 1px solid #cccccc;
+}
+
+.wrapper .profile_counts .profile_counts_wrap:first-child {
+  padding-top: 0;
+}
+
+.wrapper .profile_counts .profile_counts_wrap:last-child {
+  padding-bottom: 0px;
+  border-bottom: 0px;
+}
+
+.wrapper .profile_counts .profile_counts_wrap .item {
+  padding: 20px 10px;
+  text-align: center;
+  border-radius: 10px;
+  cursor: pointer;
+  transition: all .2s ease;
+}
+
+.wrapper .profile_counts .profile_counts_wrap .item:hover {
+  background: #e4e9ef;
+}
+
+.wrapper .profile_counts .profile_counts_wrap .item .fas {
+  font-size: 30px;
+  margin-bottom: 20px;
+}
+
+.wrapper .profile_counts .profile_counts_wrap .item .title {
+  font-size: 14px;
+}
+
+.tag {
+  background: #e4e9ef;
+  padding: 10px 15px;
+  display: inline-block;
+  border-radius: 10px;
+  margin: 5px;
+  font-size: 14px;
+}
+
+.center-button {
+  display: block; 
+  margin-left: auto;
+  margin-right: auto;
+  margin-top: 20px;
+}
+
+
+
 
 
 
@@ -135,3 +339,11 @@
     margin-top: -130px;
   }
   
+  
+
+
+
+
+ 
+
+
